Only kill if no graceful exit.

+++ dash/runner.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"strings"
 	"syscall"
+	"time"
 
 	"k8s.io/klog/v2"
 )
@@ -91,6 +92,20 @@ func (r *runner) RunE(ctx context.Context, splitResult *SplitResult,
 
 	go func() {
 		<-ctx.Done()
+
+		_ = cmd.Process.Kill()
+
+		if cmd.ProcessState.Exited() {
+			return
+		}
+
+		// Wait for 1s to exit the good way before trying to kill it.
+		time.Sleep(time.Second)
+
+		if cmd.ProcessState.Exited() {
+			return
+		}
+
 		pgid, err := syscall.Getpgid(cmd.Process.Pid)
 		if err == nil {
 			_ = syscall.Kill(-pgid, 15) // note the minus sign
@@ -103,7 +118,6 @@ func (r *runner) RunE(ctx context.Context, splitResult *SplitResult,
 		}
 		*/
 
-		_ = cmd.Process.Kill()
 		//klog.ErrorS(err, "failed to kill process the usual way")
 
 		klog.V(5).InfoS("Killed process group",
